+++ src/test/java/store/domain/store/StockTest.java
@@ -1,11 +1,9 @@
-package store.domain;
+package store.domain.store;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
 import org.junit.jupiter.api.Test;
-
-import store.domain.store.Stock;
 
 class StockTest {
